+++ src/App.js
@@ -1,14 +1,12 @@
-import { useState, useEffect, useCallback, useRef } from "react";
+import { useState, useEffect, useRef } from "react";
 import useRandomNumbers from "./useRandomNumberHook";
 import getData from "./getData";
 import getAsyncData from "./getAsyncData";
-import { mapArrayValuesToUniqueIds } from "./keys";
 import "./index.css";
 
 function App() {
-  const [data, setData] = useState(() => getData()); // lazy init
+  const [data, setData] = useState(() => getData());
   const [asyncData, setAsyncData] = useState([]);
-  // const [uniqueKeyArr, setuniqueKeyArr] = useState([]);
   const paragraphRefs = useRef([]);
 
   useEffect(() => {
@@ -24,14 +22,14 @@ function App() {
   const combinedData = [...data, ...asyncData];
   const uniqueNumbers = [...new Set(combinedData)];
   const sortedNumbers = uniqueNumbers.sort((a, b) => a - b);
-  const sortedNumberElements = Array.from(
-    mapArrayValuesToUniqueIds(sortedNumbers).entries()
-  ).map(([key, value]) => {
-    const ref = useRef(null);
-    paragraphRefs.current[key] = ref;
+  const sortedNumberElements = sortedNumbers.map((value, index) => {
     return (
-      <p key={key} className="number-paragraph" ref={ref}>
-        <span>{value}</span>
+      <p
+        key={`sorted-${index}-${value}`}
+        className="number-paragraph"
+        ref={(element) => (paragraphRefs.current[index] = element)}
+      >
+        {value}
       </p>
     );
   });
@@ -47,45 +45,50 @@ function App() {
     setRange(event.target.value);
   };
 
-  const hookArr = hookAsyncData.map((number) => {
-    const ref = useRef(null);
-    paragraphRefs.current.push(ref);
+  const hookArr = hookAsyncData.map((number, index) => {
     return (
-      <p key={number} className="number-paragraph" ref={ref}>
+      <p
+        key={`hook-${index}-${number}`}
+        className="number-paragraph"
+        ref={(element) =>
+          (paragraphRefs.current[index + sortedNumbers.length] = element)
+        }
+      >
         <span>{number}</span>
       </p>
     );
   });
 
-  //Task 5 ************************************************
   const [input, setInput] = useState(null);
   const [filteredData, setFilteredData] = useState([]);
   const allData = [...sortedNumbers, ...hookAsyncData];
-
-  function debounce(cb, delay = 1000) {
-    let timeoutId;
-    return (...args) => {
-      clearTimeout(timeoutId);
-      timeoutId = setTimeout(() => {
-        cb(...args);
-      }, delay);
-    };
-  }
 
   const filterFunction = () => {
     const filter = allData.filter((num) => num == input);
     return filter;
   };
 
-  const invokeDebouncedValidation = useCallback(
-    debounce(() => {
-      setFilteredData(filterFunction());
-    }, 500),
-    [input, allData]
-  );
+  let debounceTimeout;
+  const debounce = (cb, delay = 1000) => {
+    clearTimeout(debounceTimeout);
+    debounceTimeout = setTimeout(() => {
+      cb();
+    }, delay);
+  };
 
-  const filterMap = filteredData.map((number) => {
-    return <p key={number}>{number}</p>;
+  const invokeDebouncedValidation = () => {
+    debounce(() => {
+      const filtersss = filterFunction();
+      setFilteredData(filtersss);
+    }, 500);
+  };
+
+  useEffect(() => {
+    invokeDebouncedValidation();
+  }, [input, allData]);
+
+  const filterMap = filteredData.map((number, index) => {
+    return <p key={`filter-${index}-${number}`}>{number}</p>;
   });
 
   const initiateSearchParam = (event) => {
@@ -94,30 +97,22 @@ function App() {
   };
 
   const highlightNumbers = (inputValue) => {
-    // paragraphs = document.querySelectorAll('.number-paragraph');
-
-    // paragraph.forEach
-    paragraphRefs.current.forEach((ref) => {
-      const paragraph = ref.current;
-      const number = paragraph.querySelector("span");
-      if (number === inputValue) {
-        paragraph.classList.add("highlighted");
+    const parsedInput = parseInt(inputValue);
+    paragraphRefs.current.forEach((ref, index) => {
+      const number = allData[index];
+      if (number === parsedInput) {
+        ref.classList.add("highlighted");
       } else {
-        paragraph.classList.remove("highlighted");
+        ref.classList.remove("highlighted");
       }
     });
   };
-
-  useEffect(() => {
-    invokeDebouncedValidation();
-  });
 
   return (
     <>
       <p>
         <label>
           <input
-            ref={paragraphRefs}
             placeholder="Enter a number you want to find"
             type="number"
             value={input}
@@ -125,15 +120,10 @@ function App() {
           />
         </label>
       </p>
-      {input ? (
-        filterMap.map((number) => <p>{number}</p>)
-      ) : (
-        <p> No result found </p>
-      )}
+      {input ? filterMap : <p> No result found </p>}
       <p>************************</p>
-      {sortedNumberElements.map((number) => {
-        return <p>{number}</p>;
-      })}
+      {sortedNumberElements}
+      {hookArr}
       <p>
         <label>
           <input
@@ -151,9 +141,6 @@ function App() {
         </label>
       </p>
       <button onClick={handleClick}>Regenerate Numbers</button>
-      {hookArr.map((number) => {
-        return <p>{number}</p>;
-      })}
       {error && <p>{error}</p>}
     </>
   );
